fix: move refresh indicator outside transformed <main> so fixed positioning works
CSS quirk: when an ancestor has transform, position:fixed descendants are
positioned relative to that ancestor, not the viewport. The indicator was
inside <main> (which gets translateY during pull), so it slid with the page
instead of staying at the top of the viewport.

- Extract indicator into PullToRefreshIndicator, rendered in the layout as
  a sibling of MobileMainShell (outside the transformed <main>)
- PullToRefresh now only handles touch events and drives context state;
  it renders no UI of its own
- Indicator fixed positioning now correctly anchors to the viewport,
  appearing above the slid-down page while loading
