+++ src/app/(main)/layout.tsx
@@ -2,6 +2,7 @@ import { Suspense } from "react";
 import { Sidebar, MobileNav } from "@/components/layout/sidebar";
 import { MobileHeader } from "@/components/layout/mobile-header";
 import { MobileMainShell } from "@/components/layout/mobile-main-shell";
+import { PullToRefreshIndicator } from "@/components/layout/pull-to-refresh-indicator";
 import { PrivacyModeProvider } from "@/components/layout/privacy-mode-context";
 import { PullToRefreshProvider } from "@/components/layout/pull-to-refresh-context";
 import { getSession } from "@/lib/auth-session";
@@ -27,6 +28,7 @@ export default function MainLayout({
         <Suspense fallback={<Sidebar userImage={null} userName={null} />}>
           <SidebarWithSession />
         </Suspense>
+        <PullToRefreshIndicator />
         <MobileMainShell>
           <MobileHeader />
           <div className="mx-auto w-full max-w-6xl p-4 md:p-6">{children}</div>
